requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
onnx>=1.14.0
onnxruntime>=1.15.0
onnxsim>=0.4.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.10.0
numpy>=1.24.0

# Training utilities
tensorboard>=2.13.0
tqdm>=4.65.0
pyyaml>=6.0

# Data handling
pandas>=2.0.0
scikit-learn>=1.3.0

# Optional: for advanced features
matplotlib>=3.7.0
pillow>=10.0.0

